refactor(libs): ecs is now top level in src, adding tsconfig shortcuts
