+++ src/app/ecs/collider.ts
@@ -1,7 +1,7 @@
-import { EcsComponent } from './component';
-import { EcsEntity } from './entity';
-import { EcsScene } from './scene';
-import { EcsTransform } from './transform';
+import { EcsComponent } from './ecs-component';
+import { EcsEntity } from './ecs-entity';
+import { EcsScene } from './ecs-scene';
+import { EcsTransform } from './ecs-transform';
 import { Rectangle, Shape } from './geometry';
 
 export class Collider extends EcsComponent {
