+++ src/app/ecs/ecs-object.ts
