+++ src/app/ecs/input.ts
