--- src/app/lib/ecs/index.ts
@@ -1,10 +1,0 @@
-export * from './camera';
-export * from './collider';
-export * from './component';
-export * from './debug';
-export * from './entity';
-export * from './geometry';
-export * from './input';
-export * from './renderer';
-export * from './scene';
-export * from './transform';
